refactor: separate checkpoint events and durable progress evidence
