+++ .agents/scripts/tests/test-pr-checkpoint-revision.py
@@ -16,6 +16,7 @@ import tempfile
 import unittest
 
 SCRIPTS = pathlib.Path(__file__).resolve().parents[1]
+sys.path.insert(0, str(SCRIPTS))
 
 
 def load(name, filename):
